Type safety

+++ src/com/asdev/libjam/mt/SmartTouch.java
@@ -11,7 +11,7 @@ public class SmartTouch implements Runnable {
 	public void run() {
 		while(true){
 			
-			int execS = System.nanoTime();
+			long execS = System.nanoTime();
 			for(int i = 0; i < 10; i ++){
 				long val = TouchHandler.syncTTGet(i);
 				if(Math.abs(execS - val) >= TOUCH_TIMEOUT_NS){
